requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.6.0
django-extensions==4.1
djangorestframework==3.16.1
openpyxl==3.1.5
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
